requirements.txt
# Core Dependencies
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0

# NVIDIA Model Optimizer
nvidia-modelopt>=0.13.0
# Note: Install with: pip install "nvidia-modelopt[torch]" --extra-index-url https://pypi.nvidia.com

# Transformers and Model Loading
transformers>=4.40.0
accelerate>=0.27.0
huggingface-hub>=0.20.0
safetensors>=0.4.0

# LIBERO Benchmark
# Note: Install from source - https://github.com/Lifelong-Robot-Learning/LIBERO
# git clone https://github.com/Lifelong-Robot-Learning/LIBERO.git
# cd LIBERO && pip install -e .

# Robotics Dependencies (for LIBERO)
gym>=0.26.0
robosuite>=1.4.0
# Note: May require MuJoCo installation

# Vision and Language Models
qwen-vl-utils  # For Qwen2.5-VL model utilities
einops>=0.7.0
timm>=0.9.0

# Model Export
onnx>=1.15.0
onnxruntime-gpu>=1.17.0
# tensorrt>=8.6.0  # Optional: TensorRT for deployment
# polygraphy  # Optional: TensorRT debugging utilities

# Data Handling
datasets>=2.16.0
h5py>=3.10.0
opencv-python>=4.9.0

# Configuration and Logging
pyyaml>=6.0
tensorboard>=2.15.0
wandb>=0.16.0  # Optional: For experiment tracking
tqdm>=4.66.0
rich>=13.7.0

# Evaluation and Metrics
scipy>=1.11.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0

# Utilities
click>=8.1.0  # For CLI scripts
python-dotenv>=1.0.0
psutil>=5.9.0  # For memory profiling
